+++ DIS/server/routes/api/sign/index.js
@@ -4,8 +4,10 @@ const jwt = require('jsonwebtoken');
 const crypto = require('crypto');
 const request = require('request');
 
-router.post('/signin', (req, res, next) => {
+router.post('/login', (req, res, next) => {
     res.send({ success: true });
 });
-
+router.get('/sign-up', function(req, res, next) {
+    console.log(req.body)
+  });
 module.exports = router;
