--- DIS/server/routes/login.js
@@ -1,9 +1,0 @@
-var express = require('express');
-var router = express.Router();
-
-/* 로그인 */
-router.get('/login', function(req, res, next) {
-  console.log(req.body)
-});
-
-module.exports = router;
